requirements.txt
# Core Dependencies
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0

# FastAPI and Web Framework
fastapi==0.115.12
starlette==0.46.2
uvicorn==0.34.2
python-multipart==0.0.20
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
websockets==15.0.1

# Data Processing
pandas==2.3.0
numpy==2.3.0
matplotlib==3.10.3
pillow==11.2.1
scipy==1.16.0

# Machine Learning
scikit-learn==1.7.0
joblib==1.5.1
sentence-transformers==4.1.0
transformers==4.52.4
torch==2.7.1
tokenizers==0.21.2
safetensors==0.5.3
huggingface-hub==0.33.0

# LLM API
groq==0.28.0
openai==1.76.0

# Google AI (if needed)
google-genai==1.12.1
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0

# Configuration and Environment
python-dotenv==1.1.0
pydantic==2.11.3
pydantic_core==2.33.1
pydantic-settings==2.8.2

# Utilities
filelock==3.18.0
fsspec==2025.5.1
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0

# Image Processing
opencv-python==4.11.0.86
fonttools==4.58.4
kiwisolver==1.4.8
mpmath==1.3.0
networkx==3.5

# Protocol Buffers and RPC
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2

# Security
rsa==4.9.1
httplib2==0.22.0
uritemplate==4.1.1
idna==3.10
jiter==0.9.0
setuptools==80.9.0

pydantic_core==2.33.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==4.1.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
